feat: extract conversation filter bar
