+++ src/components-next/common/filters/index.ts
@@ -1,3 +1,2 @@
 export * from './FilterBar';
 export * from './FilterButton';
-export * from './ConversationFilterBar';
